requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
sentence-transformers==3.0.1
tqdm==4.66.4
joblib==1.4.2
pyyaml==6.0.2
